Showcase is no more, moved tests to appropriate place
