+++ src/tools/tests/Variable.arcs
@@ -1,6 +1,12 @@
 meta
   namespace: src.tools.tests
 
+particle Variable
+  h0: reads ~a
+  h1: writes ~a
+  h2: reads ~b with {num: Number}
+  h3: writes ~b with {flavor: Text}
+
 particle Ingest
   action: writes [Action {sessionId: Text, action: Text, timestampInMs: Number}]
