+++ src/tools/tests/VariableEntityTest.kt
@@ -8,7 +8,7 @@ import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 
 @RunWith(JUnit4::class)
-class GenerationBehaviorTest {
+class VariableEntityTest {
 
     @Test
     fun variable_equality() {
